+++ sonar-project.properties
@@ -1,0 +1,25 @@
+sonar.projectKey=nuvlaedge_peripheral-manager-modbus
+sonar.organization=nuvlaedge
+
+# This is the name and version displayed in the SonarCloud UI.
+#sonar.projectName=peripheral-manager-network
+#sonar.projectVersion=not provided
+
+sonar.python.version=3.8, 3.9
+sonar.python.coverage.reportPaths=coverage.xml
+
+# Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
+sonar.sources=code/
+#sonar.exclusions=
+#sonar.inclusions=
+
+# Path to tests
+#sonar.tests=
+#sonar.test.exclusions=
+#sonar.test.inclusions=
+
+# Encoding of the source code. Default is default system encoding
+#sonar.sourceEncoding=UTF-8
+
+# Exclusions for copy-paste detection
+#sonar.cpd.exclusions=
