+++ tox.ini
@@ -1,0 +1,7 @@
+[tox]
+envlist = py38
+skipsdist = True
+
+[testenv]
+deps = -rcode/requirements.tests.txt
+commands = pytest --junitxml=test-report.xml --cov=code/modbus --cov-report=xml --cov-config=code/.coveragerc --cov-branch
